+++ builtins/es2015/Uint32Array.ts
