+++ builtins/es2015/typedarrays.ts
@@ -6,7 +6,7 @@ import _Uint8ClampedArray from './Uint8ClampedArray';
 import _Int16Array from './Int16Array';
 import _Uint16Array from './Uint16Array';
 import _Int32Array from './Int32Array';
-import _UInt32Array from './UInt32Array';
+import _Uint32Array from './Uint32Array';
 import _Float32Array from './Float32Array';
 import _Float64Array from './Float64Array';
 
@@ -19,7 +19,7 @@ export default () => ({
   Int16Array: _Int16Array(),
   Uint16Array: _Uint16Array(),
   Int32Array: _Int32Array(),
-  UInt32Array: _UInt32Array(),
+  Uint32Array: _Uint32Array(),
   Float32Array: _Float32Array(),
   Float64Array: _Float64Array()
 });
